### 2018-11-20 #1093 @arfon
:wave: @emchristensen - could you please rename your `paper.Rmd` file to `paper.md` please?
