### 2018-11-20 f2938d7
Deploy weecology/portalr to github.com/weecology/portalr.git:gh-pages
