### 2018-11-20 d8a60cd
removed chunk of code leftover from when it was a .Rmd
