### 2018-11-20 #1093 @emchristensen
I just removed a bit of code leftover when it was a .Rmd, otherwise looks good.
