### 2018-11-20 c01aec2
Deploy weecology/portalr to github.com/weecology/portalr.git:gh-pages
