+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: '2.2'
-pkgdown: 1.1.0
+pkgdown: 1.2.0
 pkgdown_sha: ~
 articles:
   portal_researcher_examples: portal_researcher_examples.html
