### 2018-11-26 #1098 @whedon
[REVIEW]: portalr: an R package for summarizing and using the Portal Project Data
**Submitting author:** @emchristensen (<a href="http://orcid.org/0000-0002-5635-2502">Erica Christensen</a>)
**Repository:** <a href="https://github.com/weecology/portalr" target ="_blank">https://github.com/weecology/portalr</a>
**Version:** v0.1.4
