### 2018-12-12 c1f34c0
Use from_zenodo argument
