+++ tests/testthat/test-01-data-retrieval.R
@@ -4,11 +4,11 @@ portal_data_path <- tempdir()
 
 test_that("download_observations and check_for_newer_data work", {
   skip_on_cran() # these download checks take a while to run
-  expect_error(download_observations(portal_data_path, version = "1.20.0"), NA)
+  expect_error(download_observations(portal_data_path, version = "1.20.0", from_zenodo = FALSE), NA)
   expect_true(check_for_newer_data(portal_data_path))
   unlink(file.path(portal_data_path, "PortalData"), recursive = TRUE)
 
-  expect_error(download_observations(portal_data_path, version = "1.5"), NA)
+  expect_error(download_observations(portal_data_path, version = "1.6", from_zenodo = FALSE), NA)
   expect_true(check_for_newer_data(portal_data_path))
   unlink(file.path(portal_data_path, "PortalData"), recursive = TRUE)
 
