### 2018-12-12 290372f
fix get_github_releases
Previous code would not necessarily get the last page of github releases - this fixes that by adding a check to keep retrieving the info from github if the link is still "next" (more results) or "last" (last page of results).
