### 2018-12-12 0c4bdf2
Update DESCRIPTION
Use https://weecology.github.io/portalr/
Fix link to portaldata
