### 2018-12-12 ba40781
Deploy weecology/portalr to github.com/weecology/portalr.git:gh-pages
