### 2018-12-12 eed25d3
Update DESCRIPTION
