### 2018-12-12 04d34b5
Disable option to download from Zenodo
Just until Zenodo offers API to get versions
