+++ R/GetData.R
@@ -27,7 +27,6 @@ full_path <- function(reference_path, base_path = getwd()) {
 #'   TODO: incorporate data retriever into this when it's pointed at the github repo
 #' @param base_folder Folder into which data will be downloaded
 #' @param version Version of the data to download (default = "latest")
-#' @param from_zenodo logical; if `TRUE`, get info from Zenodo, otherwise GitHub
 #'
 #' @return None
 #'
@@ -38,10 +37,10 @@ full_path <- function(reference_path, base_path = getwd()) {
 #' }
 #'
 #' @export
-download_observations <- function(base_folder = "~", version = "latest", from_zenodo = TRUE)
+download_observations <- function(base_folder = "~", version = "latest")
 {
   # get version info
-  releases <- get_data_versions(from_zenodo, halt_on_error = TRUE)
+  releases <- get_data_versions(from_zenodo = FALSE, halt_on_error = TRUE)
 
   # match version
   if (version == "latest")
@@ -107,26 +106,27 @@ download_observations <- function(base_folder = "~", version = "latest", from_ze
 #'   `zipball_url` (download URLs for the corresponding zipped release).
 #'
 #' @export
-get_data_versions <- function(from_zenodo = TRUE, halt_on_error = FALSE)
+get_data_versions <- function(from_zenodo = FALSE, halt_on_error = FALSE)
 {
-  releases <- tryCatch(
-    {
-      if (from_zenodo)
-      {
-        get_zenodo_latest_release()
-      } else {
+  releases   <- tryCatch(
+    # {
+    #   if (from_zenodo)
+    #   {
+    #     get_zenodo_latest_release()
+    #   } else {
         get_github_releases()
-      }
-    },
+    #   }
+    # }
+    ,
     error = function(e) {
       if (halt_on_error) {
         stop(e)
       } else {
         e
       }
-    },
+     },
     warning = function(w) w
-  )
+   )
   if (!is.data.frame(releases))
   {
     return(NULL)
