### 2018-12-12 9a01247
Download from repo (#175)
* Disable option to download from Zenodo

Just until Zenodo offers API to get versions
