+++ README.md
@@ -77,19 +77,19 @@ browseVignettes("portalr")
 
 ## More Information
 
-#### [Portal Data Repo](github.com/weecology/PortalData)
+#### [Portal Data Repo](https://github.com/weecology/PortalData)
 
 The data repo contains useful details for issues with data collection,
 and background on why we handle them the way we do. Of course, it also
 contains the raw data, if you would like to create more complex data
 summaries than what is provided here.
 
-#### [The Portal Project](portal.weecology.org/)
+#### [The Portal Project](https://portal.weecology.org/)
 
 Find a list of previous publications using the Portal data at our
 website.
 
-#### [The Portal Blog](portalproject.wordpress.com/)
+#### [The Portal Blog](https://portalproject.wordpress.com/)
 
 Follow our blog to get the latest news on what is happening with our
 project and at the site.
