### 2018-12-21 fa2c72d
Update README.Rmd
fix typo in .Rmd version of readme
