### 2018-12-26 666d316
Fix calculation of abundance per species
I think the current code in the vignette is not working as intended (see Fig https://weecology.github.io/portalr/articles/rodent-abundance-demo.html#figure-abundance-over-time).
I tried to fix with a couple of minor changes.
An alternative would be to call `rodent_abundance_by_plot <- abundance(path = portal_data_path, time = "date", level = "plot", shape = "flat") ` so there would be no need to call `gather`, just `count` to summarise abundance per species
