### 2018-12-27 weecology/portalr#179 @Pakillo
JOSS review: add DOI for papers, and revise R citation
Hi there,

I'm reviewing `portalr` for JOSS (https://github.com/openjournals/joss-reviews/issues/1098). I have been playing with the package and it works nicely!

I only have a few minor comments or suggested changes that I will leave as issues here. I have already submitted a couple of pull requests with very minor fixes for one vignette.

So, first, I should check that "all archival references that should have a DOI list one (e.g., papers, datasets, software)?". Could you please include the DOI for the papers currently missing it in the reference list?

Also, the correct citation for R, as suggested by `citation`, would be:
R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

Thanks!
