### 2018-12-27 weecology/portalr#180 @Pakillo
JOSS review: Consider renaming/restructuring load_ functions?
Hi,

Another comment related to my JOSS review (https://github.com/openjournals/joss-reviews/issues/1098).

As my first encounter with the package, I got a bit confused at first with the load_* functions (https://weecology.github.io/portalr/reference/load_data.html). It isn't obvious from their names what particular data these functions are loading. As there is `load_plant_data` and `load_ant_data`, maybe `load_data` could be renamed as `load_rodent_data` for consistency, or `load_trapping_data` to account for the fact that not only rodents fall in traps, but also reptiles, birds, etc?

I note there is already a `load_trapping_data` function, but this one does not seem to load any abundance or catch data like the others, but rather the data on when and where trappings took place? Then this function role would be different from the other three (i.e. loading data from rodents (mostly), plants, and ants), and could be perhaps renamed as an independent function? (e.g. as the others you have to get weather data, NDVI, etc).

Just a suggestion... Because it took me a couple of minutes to learn what each function would do. But up to you of course!
