### 2018-12-27 weecology/portalr#181 @Pakillo
Add more badges?
Hi,

Another suggestion related to my JOSS review (https://github.com/openjournals/joss-reviews/issues/1098): I would recommend to add a few more badges to the readme in the GitHub repo, to give some clear signals of the package status. For instance, one CRAN badge (can show number of downloads too if you wish), and one badge showing test coverage. As you have clearly worked hard on testing your functions, I think it would be good to show that clearly in the README. At least test coverage is one of the quality indicators I seek when examining new packages!

Again, just a suggestion. Up to you if you prefer not to!
