### 2018-12-27 #1098 @Pakillo
Hi there,

Thanks for inviting me to review `portalr`. I have been playing with the package and it works nicely! The work you are doing to manage and facilitate reuse and sharing of data from the Portal Project is really impressive. Leading by example.

I only had a few minor comments or suggested changes.

I have opened a few issues in the original repo (https://github.com/weecology/portalr/issues/created_by/Pakillo), most of them just suggestions to consider, and have submitted a couple of pull requests (https://github.com/weecology/portalr/pull/177 & https://github.com/weecology/portalr/pull/178) with very minor fixes for one vignette.

I think some functions would benefit from a somewhat more developed documentation (e.g. specify output values, add examples, improve descriptions of expected parameters...). But I understand that takes time, and at least all the functions I tried seem to work well.

The only item I could not yet check from the review checklist above is that of having DOIs for all references. I have requested them to be added (https://github.com/weecology/portalr/issues/179). Once this is done I am happy to recommend publication!
