### 2018-12-28 4fba8d5
add CRAN and code coverage badges
