### 2018-12-28 weecology/portalr#180 @ha0ye
This partly historical legacy of the package originally being developed around the rodent data, and then adding in the functionality for plant and ant data later.

Ideally, most users shouldn't be using the `load_*` functions at all, but some folks who have built on older versions of the package may rely on `load_data`. The solution is probably some form of:
* add `load_rodent_data` as an alias, and deprecate `load_data`
* add 1 or 2 functions that just load in the metadata tables, to replace some uses of `load_data` (as in the `rodent-abundance-demo` vignette)
