### 2019-01-02 498f84d
added doi to references in paper.md; updated R ref
