+++ paper/paper.bib
@@ -4,7 +4,8 @@
     journal={Ecology},
     volume={75},
     year={1994},
-    pages={438-445}
+    pages={438-445},
+    doi={10.2307/1939547}
 }
 
 @article{White:2018,
@@ -13,13 +14,16 @@
     journal={Methods in Ecology and Evolution},
     volume={},
     year={2018},
-    pages={}
+    pages={},
+    doi={10.1111/2041-210X.13104}
 }
 
-@article{RCoreTeam2016,
+@article{RCoreTeam2018,
     author={{R Core Team}},
-    year={2016},
-    title={R: a language and environment for statistical computing}
+    year={2018},
+    title={R: a language and environment for statistical computing},
+    url={https://www.R-project.org/},
+    address={Vienna, Austria}
 }
     
 @article{Ernest:2018b,
