### 2019-01-07 weecology/portalr#179 @emchristensen
fixed in #183
