### 2019-01-08 be4a480
Various tibble-related fixes - make regression test pipes consistent - use tibble::as_tibble instead of dplyr version - fix regression tests (related to change in tibble functionality) - adding ungroup() to processing chains so that results are not grouped tibbles
