+++ tests/testthat/test-99-regression.R
@@ -5,58 +5,58 @@ portal_data_path <- tempdir()
 test_that("data generated by default setting is same", {
   data <- abundance(portal_data_path, level = 'Site',
                     type = "Rodents", plots = "all", unknowns = FALSE,
-                    min_plots = 24, shape = "crosstab", time = "period")
-  data <- dplyr::filter(data, period < 434)
+                    min_plots = 24, shape = "crosstab", time = "period") %>%
+    dplyr::filter(period < 434)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
   expect_identical(digest::digest(data), "f7365a1b64bd3aa30579abd37c6de863")
 })
 
 test_that("data generated by level = treatment, plots = longterm is same", {
-  data <- abundance(portal_data_path, level = 'treatment', plots = "longterm")
-  data <- dplyr::filter(data, period < 434)
+  data <- abundance(portal_data_path, level = 'treatment', plots = "longterm") %>%
+    dplyr::filter(period < 434)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
   expect_identical(digest::digest(data), "5516d861d63a1305ff4b39b62c29e5a1")
 })
 
 test_that("data generated by level = plot, time = newmoon, type = granivore, shape = flat is same", {
   data <- abundance(portal_data_path, level = 'plot', type = "granivores",
-                    shape = "flat", time = "newmoon")
-  data <- dplyr::filter(data, newmoonnumber < 465)
+                    shape = "flat", time = "newmoon") %>%
+    dplyr::filter(newmoonnumber < 465)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
   expect_identical(digest::digest(data), "22f018525c242edaa3f94814e04d230b")
 })
 
 test_that("data generated by unknowns = T, min_plots = 1 is same", {
-  data <- abundance(portal_data_path, min_plots = 1, unknowns = TRUE)
-  data <- dplyr::filter(data, period < 434)
-  data <- dplyr::select(data, period, BA, DM, DO, DS, "NA", OL, OT, other, PB,
-                      PE, PF, PH, PI, PL, PM, PP, RF, RM, RO, SF, SH, SO)
+  data <- abundance(portal_data_path, min_plots = 1, unknowns = TRUE) %>%
+    dplyr::filter(period < 434) %>%
+    dplyr::select(period, BA, DM, DO, DS, "NA", OL, OT, other, PB,
+                  PE, PF, PH, PI, PL, PM, PP, RF, RM, RO, SF, SH, SO)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
   expect_identical(digest::digest(data), "716cba45b04466dea49bccd2844ef5da")
 })
 
 test_that("data generated by plots = c(4, 8, 10, 12) is same", {
   data <- get_rodent_data(path = portal_data_path, plots = c(4, 8, 10, 12),
-                  na_drop = TRUE, zero_drop = FALSE)
-  data <- dplyr::filter(data, period < 450)
+                          na_drop = TRUE, zero_drop = FALSE) %>%
+    dplyr::filter( period < 450)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
   expect_identical(digest::digest(data), "c4eb4083894c3790ceeba6b8ef8d3579")
 })
 
 test_that("biomass data generated by level = plot is same", {
-  data <- biomass(portal_data_path, level = "plot")
-  data <- dplyr::filter(data, period < 434)
+  data <- biomass(portal_data_path, level = "plot") %>%
+    dplyr::filter(period < 434)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
 
[…]
