### 2019-01-10 weecology/portalr#180 @ha0ye
Proposed changes:
* rename `get_***_data()` --> `summarize_***_data()`
* rename `load_data()` --> `load_rodent_data()`
* simplify and export `load_generic_data()` for loading individual table files (e.g. for specific metadata files)
