### 2019-01-10 weecology/portalr#180 @ha0ye
@diazrenata Did you have thoughts on renaming `get_stake_data()`?
