### 2019-01-10 weecology/portalr#180 @diazrenata
It seems more intuitive to me to refer to it as `get_individual_data`, but I can see the rationale for `get_stake_data`.
