### 2019-01-10 weecology/portalr#180 @ha0ye
@diazrenata Well, since we're renaming `get_rodent_data()` to `summarize_rodent_data()`, are you in favor of `summarize_individual_data()`? I was wondering if `summarize_individual_rodents()` would be clearer about it being the rodent data.
