### 2019-01-10 1622863
rename functions: * `get_***_data()` -> `summarize_***_data()`
