+++ tests/testthat/test-99-regression.R
@@ -36,7 +36,7 @@ test_that("data generated by unknowns = T, min_plots = 1 is same", {
 })
 
 test_that("data generated by plots = c(4, 8, 10, 12) is same", {
-  data <- get_rodent_data(path = portal_data_path, plots = c(4, 8, 10, 12),
+  data <- summarize_rodent_data(path = portal_data_path, plots = c(4, 8, 10, 12),
                           na_drop = TRUE, zero_drop = FALSE) %>%
     dplyr::filter( period < 450)
   attributes(data) <- attributes(data)[sort(names(attributes(data)))]
