+++ R/GetData.R
@@ -1,539 +1,0 @@
-#' @importFrom graphics plot
-#' @importFrom stats aggregate
-#' @importFrom utils download.file read.csv unzip read.table tail
-
-#' @title Full Path
-#' @description Return normalized path for all operating systems
-#' @param reference_path a path to join with current working directory
-#' @param base_path Current working directory else path given
-#'
-#' @return Full path
-#'
-#' @examples
-#' full_path('PortalData/Rodents/Portal_rodent.csv')
-#' full_path('PortalData/Rodents/Portal_rodent.csv', '~')
-#'
-#' @noRd
-full_path <- function(reference_path, base_path = getwd()) {
-  base_path <- normalizePath(base_path)
-  path <- normalizePath(file.path(base_path, reference_path), mustWork = FALSE)
-  return(path)
-}
-
-#' @title Download the PortalData repo
-#'
-#' @description This downloads the latest portal data regardless if they are
-#'   actually updated or not.
-#'   TODO: incorporate data retriever into this when it's pointed at the github repo
-#' @param base_folder Folder into which data will be downloaded
-#' @param version Version of the data to download (default = "latest")
-#'
-#' @return None
-#'
-#' @examples
-#' \donttest{
-#'   download_observations()
-#'   download_observations("~/old-data", version = "1.50.0")
-#' }
-#'
-#' @export
-download_observations <- function(base_folder = "~", version = "latest")
-{
-  # get version info
-  releases <- get_data_versions(from_zenodo = FALSE, halt_on_error = TRUE)
-
-  # match version
-  if (version == "latest")
-  {
-    match_idx <- 1
-  } else {
-    # Normalize version number
-    if (grepl("^[0-9]+\\.[0-9]+$", version))
-    {
-      version <- paste0(version, ".0")
-    }
-    if (!grepl("^[0-9]+\\.[0-9]+\\.0$", version))
-    {
-      stop("Invalid version number; given, ", version, call. = FALSE)
-    }
-
-    match_idx <- match(version, releases$version)
-    if (length(match_idx) != 1 || is.na(match_idx))
-    {
-      stop("Did not find a version of the data matching, ", version, call. = FALSE)
-    }
-  }
-
-  # Attemt to download the zip file
-  message("Downloading version ", releases$version[match_idx], " of the data...")
-  zip_download_path <- releases$zipball_url[match_idx]
-  zip_download_dest <- full_path("PortalData.zip", tempdir())
-  download.file(zip_download_path, zip_download_dest, quiet = TRUE, mode = "wb")
-
-  final_data_folder <- full_path("PortalData", base_folder)
-
-  # Clear out the old files in the data folder without doing potentially dangerous
-  # recursive deleting.
-  if (file.exists(final_data_folder)) {
-    old_files <- list.files(
-      final_data_folder,
-      full.names = TRUE,
-      all.files = TRUE,
-      recursive = TRUE,
-      include.dirs = FALSE
-    )
-    file.remove(normalizePath(old_files))
-    unlink(final_data_folder, recursive = TRUE)
-  }
-
-  #Github serves this up with the -master extension. Unzip and rename to remove that.
-  primary_data_folder <- unzip(zip_download_dest, list = TRUE)$Name[1]
-  unzip(zip_download_dest, exdir = base_folder)
-  Sys.sleep(10)
-  file.remove(zip_download_dest)
-  file.rename(full_path(primary_data_folder, base_folder), final_data_folder)
-}
-
-#' @title get version and download info for PortalData
-#'
-#' @description Check either Zenodo or GitHub for the version and download link
-#'   for PortalData.
-#'
-#' @param from_zenodo logical; if `TRUE`, get info from Zenodo, otherwise GitHub
[…]
