### 2019-01-10 weecology/portalr#180 @gmyenni
The initial idea was that `stake_data` was the more or less 'raw' individual data, and `individual_data` would be the 'cleaned' individual data (where unique tags actually refer to distinct individuals), which is still on the to-do list. But we could also call this the `individual_data`, and call the cleaned data `tag_data`. That's probably more intuitive all the way around, yeah? ie `summarize_individual_rodents()` and `summarize_rodent_tags()`
