### 2019-01-10 weecology/portalr#180 @gmyenni
Note though that this function doesn't (at the moment) actually summarize. It just cleans a bit. So while that would be consistent with the 'type' of function it is, it isn't technically true.
