### 2019-01-10 8646f23
finish renaming/refactoring and update version #
