+++ R/load_data.R
@@ -35,44 +35,40 @@
 #'
 load_rodent_data <- function(path = "~", download_if_missing = TRUE, clean = TRUE)
 {
-  # set up files and NA options
-  data_files <- c("rodent_data" = file.path("Rodents", "Portal_rodent.csv"),
-                  "species_table" = file.path("Rodents", "Portal_rodent_species.csv"),
-                  "trapping_table" = file.path("Rodents", "Portal_rodent_trapping.csv"),
-                  "newmoons_table" = file.path("Rodents", "moon_dates.csv"),
-                  "plots_table" = file.path("SiteandMethods", "Portal_plots.csv"))
-  na_strings <- list(c(""), c(""), c("NA"), c("NA"), c("NA"))
-
-  # retrieve data
-  data_tables <- load_generic_data(data_files, na_strings, path, download_if_missing)
+  rodent_data <- load_datafile(file.path("Rodents", "Portal_rodent.csv"),
+                               na.strings = "", path, download_if_missing, version_message = TRUE)
+  species_table <- load_datafile(file.path("Rodents", "Portal_rodent_species.csv"),
+                                 na.strings = "", path, download_if_missing)
+  trapping_table <- load_datafile(file.path("Rodents", "Portal_rodent_trapping.csv"),
+                                 na.strings = "NA", path, download_if_missing)
+  newmoons_table <- load_datafile(file.path("Rodents", "moon_dates.csv"),
+                                 na.strings = "NA", path, download_if_missing)
+  plots_table <- load_datafile(file.path("SiteandMethods", "Portal_plots.csv"),
+                               na.strings = "NA", path, download_if_missing)
 
   # reformat species columns
-  if (!"species" %in% names(data_tables$species_table))
+  if (!"species" %in% names(species_table))
   {
-    data_tables$species_table <- dplyr::rename(data_tables$species_table,
-                                               species = speciescode)
+    species_table <- dplyr::rename(species_table, species = speciescode)
   }
 
   # convert rodent tags to characters if not already
-  data_tables$rodent_data$tag <- as.character(data_tables$rodent_data$tag)
+  rodent_data$tag <- as.character(rodent_data$tag)
 
   # remove data still under quality control
   if (clean)
   {
-    data_tables$rodent_data <- clean_data(data_tables$rodent_data,
-                                          data_tables$trapping_table,
-                                          by = c("month", "day", "year", "period", "plot"))
-    data_tables$newmoons_table <- clean_data(data_tables$newmoons_table,
-                                             data_tables$trapping_table,
-                                             by = "period")
-    data_tables$plots_table <- clean_data(data_tables$plots_table,
-                                          data_tables$trapping_table,
-                                          by = c("year", "month", "plot"))
-    data_tables$trapping_table <- dplyr::filter(data_tables$trapping_table,
-                                                qcflag == 1)
+    rodent_data <- clean_data(rodent_data, trapping_table,
+                              by = c("month", "day", "year", "period", "plot"))
+    newmoons_table <- clean_data(newmoons_table, trapping_table,
+                                 by = "period")
+    plots_table <- clean_data(plots_table, trapping_table,
+                              by = c("year", "month", "plot"))
+    trapping_table <- dplyr::filter(trapping_table, qcflag == 1)
   }
 
-  return(data_tables)
+  return(mget(c("rodent_data", "species_table", "trapping_table",
+                "newmoons_table", "plots_table")))
 }
 
 #' @rdname load_rodent_data
@@ -99,28 +95,32 @@ load_rodent_data <- function(path = "~", download_if_missing = TRUE, clean = TRU
 
[…]
