+++ man/summarize_individual_rodents.Rd
@@ -1,19 +1,19 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/summarize_individual_rodents.R
-\name{summarize_stake_data}
-\alias{summarize_stake_data}
-\alias{summarise_stake_data}
+\name{summarize_individual_rodents}
+\alias{summarize_individual_rodents}
+\alias{summarise_individual_rodents}
 \title{Return cleaned Portal rodent individual data}
 \usage{
-summarize_stake_data(path = "~", clean = TRUE, type = "Rodents",
-  length = "all", unknowns = FALSE, time = "period",
-  fillweight = FALSE, min_plots = 1, min_traps = 1,
-  download_if_missing = TRUE)
+summarize_individual_rodents(path = "~", clean = TRUE,
+  type = "Rodents", length = "all", unknowns = FALSE,
+  time = "period", fillweight = FALSE, min_plots = 1,
+  min_traps = 1, download_if_missing = TRUE)
 
-summarise_stake_data(path = "~", clean = TRUE, type = "Rodents",
-  length = "all", unknowns = FALSE, time = "period",
-  fillweight = FALSE, min_plots = 1, min_traps = 1,
-  download_if_missing = TRUE)
+summarise_individual_rodents(path = "~", clean = TRUE,
+  type = "Rodents", length = "all", unknowns = FALSE,
+  time = "period", fillweight = FALSE, min_plots = 1,
+  min_traps = 1, download_if_missing = TRUE)
 }
 \arguments{
 \item{path}{path to location of downloaded Portal data; or "repo" to
