### 2019-01-11 454cccf
change default for load_datafile and add example
