+++ man/load_datafile.Rd
@@ -4,7 +4,7 @@
 \alias{load_datafile}
 \title{read in a raw datafile from the downloaded data or the GitHub repo}
 \usage{
-load_datafile(datafile, na.strings = "NA", path = "~",
+load_datafile(datafile, na.strings = "", path = "~",
   download_if_missing = TRUE, version_message = FALSE)
 }
 \arguments{
@@ -30,4 +30,9 @@ being loaded}
 does checking for whether a particular datafile exists and then
   reads it in, using na_strings to determine what gets converted to NA. It
   can also download the dataset if it's missing locally.
+}
+\examples{
+\donttest{
+rodent_species <- load_datafile("Rodents/Portal_rodent_species.csv")
+}
 }
