### 2019-01-12 fe388d0
only access github when necessary: - only check data version on load when running interactively - enable ability to retrieve latest version info from zenodo
