### 2019-01-15 #1098 @emchristensen
Hi @Pakillo ,
Thank you for the review, and all your suggestions! We added DOIs for the references (weecology/portalr#183), and did some re-factoring of the load_functions (weecology/portalr#187) to make function names more consistent and descriptive. I can't promise we've added a significant amount of documentation as of now, but it's something we can work on as the package grows, and we will certainly keep it in mind.

Thanks again for your suggestions, and let us know if there are any other changes you would like to see.
