### 2019-01-17 #1098 @Pakillo
Hi @emchristensen et al.

Thanks for considering my suggestions. It's all fine from my side. I've completed checklist above. @karthik please let me know if there's something else I should do

Cheers
