+++ vignettes/portal_researcher_examples.Rmd
@@ -11,7 +11,8 @@ vignette: >
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  eval = Sys.getenv("portalr_vignette_build")
 )
 ```
 
