### 2019-01-22 fb136e1
using the presence of the variable
rather than the logical state of the variable itself, just using the presence of the variable, and setting the value to 1234 (rather than TRUE) to avoid confusion about how to turn it off (you have to remove it, you can't set it to FALSE)
