+++ vignettes/rodent-abundance-demo.Rmd
@@ -15,7 +15,7 @@ date_span <- list(iso_date = 0)
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  eval = Sys.getenv("portalr_vignette_build")
+  eval = nzchar(Sys.getenv("portalr_vignette_build")) 
 )
 ```
 
